Prevent BlendState mutation (immediately) after binding to GraphicsDevice
According to XNA behaviour, preventing mutation should be done with an
exception, not a debug-build-only assert. And it should be done as soon
as you set a BlendState on to a GraphicsDevice, without waiting for a
draw call.

The BlendState.GraphicsDevice property is now set in the
GraphicsDevice.BlendState setter. Previously only BlendState.DirectX set
this property in PlatformApplyState, but this is now done in a
platform-independent way.